+++ src/OpenTTY.java
@@ -22,7 +22,7 @@ public class OpenTTY extends MIDlet implements CommandListener {
     public Object shell;
     // |
     public Hashtable attributes = new Hashtable(), fs = new Hashtable(), sys = new Hashtable(), tmp = new Hashtable(), cache = new Hashtable(), cacheLua = new Hashtable(), graphics = new Hashtable(), servers = new Hashtable(), globals = new Hashtable(), userID = new Hashtable();
-    public String username = read("/home/OpenRMS", globals), build = "2026-1.18.1-03x23";
+    public String username = read("/home/OpenRMS", globals), build = "2026-1.18.1-03x24";
     // |
     // Graphics
     public Display display = Display.getDisplay(this);